# nhdb-app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NHDB - National Health Database API",
    "description": "API for managing patient records and vaccination tracking for Contoso's vaccination program",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://nhdb-app.vercel.app",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/vaccinations": {
      "get": {
        "operationId": "getVaccinationHistory",
        "summary": "Get vaccination history for a child",
        "description": "Get all vaccination records for a specific child",
        "parameters": [
          {
            "name": "childId",
            "in": "query",
            "required": true,
            "description": "The MongoDB ID of the child",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vaccination history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vaccinations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/VaccinationVisit"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "recordVaccination",
        "summary": "Record a vaccination visit",
        "description": "Record a vaccination visit for a child",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVaccinationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vaccination recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "vaccination": {
                      "$ref": "#/components/schemas/VaccinationVisit"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Vaccine already given to this child"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VaccinationVisit": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "childId": {
            "type": "string"
          },
          "visitDate": {
            "type": "string",
            "format": "date-time"
          },
          "vaccineGiven": {
            "type": "string",
            "enum": ["A", "B", "C", "none_required", "not_available"]
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateVaccinationInput": {
        "type": "object",
        "required": ["childId", "visitDate", "vaccineGiven"],
        "properties": {
          "childId": {
            "type": "string"
          },
          "visitDate": {
            "type": "string",
            "format": "date"
          },
          "vaccineGiven": {
            "type": "string",
            "enum": ["A", "B", "C", "none_required", "not_available"]
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}